Adding macros to create Figure1.png:
Figure1 is the main figure of the poster.
It consits of 5 sub images. 4 Of them are created automatically, when
running Figure1.ijm.
The 5th sub image has to be created by running Figure8.js. I was not
able to pass an argument to Figure8.js when invoking it with runMacro().